Refactor App.js for compartmentalization
