+++ sudoku-solver/src/Components/App.js
@@ -1,0 +1,12 @@
+import "./App.css";
+import { SudokuBoard } from "./SudokuBoard";
+
+function App() {
+  return (
+    <div className="App">
+      <SudokuBoard />
+    </div>
+  );
+}
+
+export default App;
